Add section to resolve objects by their path.
