+++ ftw/inflator/creation/content_creation.cfg
@@ -9,6 +9,7 @@ pipeline =
     encode-id
     constructor
     resolveuuid
+    resolvepath
     schemaupdater
     dx_schemaupdater
     workflowupdater
@@ -56,6 +57,9 @@ blueprint = collective.transmogrifier.sections.constructor
 
 [resolveuuid]
 blueprint = ftw.inflator.creation.resolveuuid
+
+[resolvepath]
+blueprint = ftw.inflator.creation.resolvepath
 
 [schemaupdater]
 blueprint = plone.app.transmogrifier.atschemaupdater
