+++ ftw/inflator/tests/interfaces.py
@@ -1,5 +1,7 @@
 from plone.directives.form import Schema
+from plone.formwidget.contenttree import ObjPathSourceBinder
 from plone.namedfile.field import NamedImage
+from z3c.relationfield.schema import RelationChoice
 from zope.interface import Interface
 
 
@@ -11,4 +13,9 @@ class IFoo(Interface):
 class IExampleDxType(Schema):
     image = NamedImage(
         title=u'Image',
+        )
+
+    relation = RelationChoice(
+        title=u'Relation',
+        source=ObjPathSourceBinder(),
         )
